requirements.txt
ansible==13.0.0
ansible-core==2.20.0
cffi==2.0.0
cryptography==46.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
resolvelib==1.2.1
six==1.17.0
tzdata==2025.2
